darkpool-client: Remove darkpool client trait
The darkpool client trait is no longer necessary as all environments use
a solidity darkpool implementation.
