+++ crates/core/Cargo.toml
@@ -5,8 +5,6 @@ edition = "2018"
 default-run = "renegade-relayer"
 
 [features]
-arbitrum = ["chain-events/arbitrum", "darkpool-client/arbitrum"]
-base = ["chain-events/base", "darkpool-client/base"]
 metered-channels = ["util/channels"]
 
 [dependencies]
