+++ crates/darkpool-client/src/client/contract_interaction.rs
@@ -1,13 +1,18 @@
 //! Defines `DarkpoolClient` helpers that allow for interacting with the
 //! darkpool contract
 
-use alloy::rpc::types::TransactionRequest;
-use alloy::{primitives::Address, rpc::types::TransactionReceipt};
+use alloy::consensus::constants::SELECTOR_LEN;
+use alloy::rpc::types::{TransactionReceipt, TransactionRequest};
+use alloy_primitives::{Address, Selector};
 use circuit_types::Nullifier;
 use circuit_types::{elgamal::EncryptionKey, fixed_point::FixedPoint, merkle::MerkleRoot};
 use constants::Scalar;
-use renegade_solidity_abi::v2::IDarkpoolV2::{DepositAuth, ObligationBundle, SettlementBundle};
+use crypto::fields::{scalar_to_u256, u256_to_scalar};
+use renegade_solidity_abi::v2::IDarkpoolV2::{
+    self, DepositAuth, DepositProofBundle, ObligationBundle, SettlementBundle,
+};
 use tracing::{info, instrument};
+use types_core::Token;
 use types_proofs::{
     IntentOnlyBoundedSettlementBundle, OrderValidityProofBundle, ValidBalanceCreateBundle,
     ValidDepositBundle,
@@ -15,11 +20,14 @@ use types_proofs::{
 use util::telemetry::helpers::backfill_trace_field;
 
 use crate::errors::DarkpoolClientError;
-use crate::traits::DarkpoolImpl;
 
-use super::DarkpoolClientInner;
+use super::DarkpoolClient;
 
-impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
+/// The set of known selectors for the Solidity darkpool
+/// TODO: Add known selectors
+const KNOWN_SELECTORS: [[u8; SELECTOR_LEN]; 0] = [];
+
+impl DarkpoolClient {
     // -----------
     // | GETTERS |
     // -----------
@@ -27,7 +35,13 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
     /// Get the current Merkle root in the contract
     #[instrument(skip_all, err)]
     pub async fn get_merkle_root(&self) -> Result<Scalar, DarkpoolClientError> {
-        self.darkpool.get_merkle_root().await
+        let depth = self.merkle_depth();
+        self.darkpool
+            .getMerkleRoot(depth)
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)
+            .map(u256_to_scalar)
     }
 
     /// Get the base fee charged by the contract
@@ -37,7 +51,12 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         in_token: Address,
         out_token: Address,
     ) -> Result<FixedPoint, DarkpoolClientError> {
-        self.darkpool.get_protocol_fee(in_token, out_token).await
+        self.darkpool
+            .getProtocolFee(in_token, out_token)
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)
+            .map(|r| r.into())
     }
 
     /// Get the external match fee charged by the contract for the given mint
@@ -46,19 +65,39 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         &self,
         mint: Address,
     ) -> Result<FixedPoint, DarkpoolClientError> {
-        self.darkpool.get_external_match_fee(mint).await
+        let usdc = Token::usdc().get_alloy_address();
+        self.darkpool
+            .getProtocolFee(mint, usdc)
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)
+            .map(|r| r.into())
     }
 
     /// Get the protocol pubkey
     #[instrument(skip_all, err)]
     pub async fn get_protocol_pubkey(&self) -> Result<EncryptionKey, DarkpoolClientError> {
-        self.darkpool.get_protocol_pubkey().await
+        let pubkey = self
+            .darkpool()
+            .getProtocolFeeKey()
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)?;
+
+        let x = u256_to_scalar(pubkey.point.x);
+        let y = u256_to_scalar(pubkey.point.y);
+        Ok(EncryptionKey { x, y })
     }
 
     /// Check whether a given root is in the contract's history
     #[instrument(skip_all, err, fields(root = %root))]
     pub async fn check_merkle_root(&self, root: MerkleRoot) -> Result<bool, DarkpoolClientError> {
-        self.darkpool.check_merkle_root(root).await
+        let root_u256 = scalar_to_u256(&root);
+        self.darkpool()
+            .rootInHistory(root_u256)
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)
     }
 
     /// Check whether a given nullifier is used
@@ -67,7 +106,21 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         &self,
         nullifier: Nullifier,
     ) -> Result<bool, DarkpoolClientError> {
-        self.darkpool.is_nullifier_spent(nullifier).await
+        let nullifier_u256 = scalar_to_u256(&nullifier);
+        self.darkpool()
+            .nullifierSpent(nullifier_u256)
+            .call()
+            .await
+            .map_err(DarkpoolClientError::contract_interaction)
+    }
+
+    /// Whether a given selector is known to the darkpool implementation
+    ///
+    /// This method is used in share recovery to decide whether a top level
+    /// trace calls the darkpool directly. If not, the relayer should trace the
+    /// transaction's calls and parse shares from darkpool subcalls.
+    pub fn is_known_selector(selector: Selector) -> bool {
+        KNOWN_SELECTORS.contains(&selector.0)
     }
 
     // -----------
@@ -86,10 +139,21 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         auth: DepositAuth,
         proof_bundle: ValidBalanceCreateBundle,
     ) -> Result<TransactionReceipt, DarkpoolClientError> {
-        let receipt = self.darkpool.create_balance(auth, proof_bundle).await?;
+        let bundle = proof_bundle.into_inner();
+        let contract_statement: IDarkpoolV2::ValidBalanceCreateStatement = bundle.statement.into();
+        let proof = bundle.proof.into();
+
+        let calldata_bundle = IDarkpoolV2::NewBalanceDepositProofBundle {
+            merkleDepth: self.merkle_depth(),
+            statement: contract_statement,
+            proof,
+        };
+        let tx = self.darkpool().depositNewBalance(auth, calldata_bundle);
+        let receipt = self.send_tx(tx).await?;
+
         let tx_hash = format!("{:#x}", receipt.transaction_hash);
         backfill_trace_field("tx_hash", &tx_hash);
-        info!("`create_balance` tx hash: {}", tx_hash);
+        info!("`create_balance` tx hash: {tx_hash}");
 
         Ok(receipt)
     }
@@ -106,10 +170,21 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         auth: DepositAuth,
         proof_bundle: ValidDepositBundle,
     ) -> Result<TransactionReceipt, DarkpoolClientError> {
-        let receipt = self.darkpool.deposit(auth, proof_bundle).await?;
+        let bundle = proof_bundle.into_inner();
+        let contract_statement: IDarkpoolV2::ValidDepositStatement = bundle.statement.into();
+        let proof = bundle.proof.into();
+
+        let calldata_bundle = DepositProofBundle {
+            merkleDepth: self.merkle_depth(),
+            statement: contract_statement,
+            proof,
+        };
+        let tx = self.darkpool().deposit(auth, calldata_bundle);
+        let receipt = self.send_tx(tx).await?;
+
         let tx_hash = format!("{:#x}", receipt.transaction_hash);
         backfill_trace_field("tx_hash", &tx_hash);
-        info!("`deposit` tx hash: {}", tx_hash);
+        info!("`deposit` tx hash: {tx_hash}");
 
         Ok(receipt)
     }
@@ -121,14 +196,13 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
         settlement_bundle0: SettlementBundle,
         settlement_bundle1: SettlementBundle,
     ) -> Result<TransactionReceipt, DarkpoolClientError> {
-        let receipt = self
-            .darkpool
-            .settle_match(obligation_bundle, settlement_bundle0, settlement_bundle1)
-            .await?;
+        let tx =
+            self.darkpool().settleMatch(obligation_bundle, settlement_bundle0, settlement_bundle1);
+        let receipt = self.send_tx(tx).await?;
 
         let tx_hash = format!("{:#x}", receipt.transaction_hash);
         backfill_trace_field("tx_hash", &tx_hash);
-        info!("`settle_match` tx hash: {}", tx_hash);
+        info!("`settle_match` tx hash: {tx_hash}");
 
         Ok(receipt)
     }
@@ -137,17 +211,25 @@ impl<D: DarkpoolImpl> DarkpoolClientInner<D> {
     // | Calldata Gen |
     // ----------------
 
-    /// Generate calldata for a `processAtomicMatchSEttle` call
+    /// Generate calldata for a `processAtomicMatchSettle` call
     pub fn gen_atomic_match_settle_calldata(
         &self,
-        receiver_address: Option<Address>,
-        internal_party_validity_proofs: &OrderValidityProofBundle,
-        match_atomic_bundle: IntentOnlyBoundedSettlementBundle,
+        _receiver_address: Option<Address>,
+        _internal_party_validity_proofs: &OrderValidityProofBundle,
+        _match_atomic_bundle: IntentOnlyBoundedSettlementBundle,
     ) -> Result<TransactionRequest, DarkpoolClientError> {
-        self.darkpool.gen_atomic_match_settle_calldata(
-            receiver_address,
-            internal_party_validity_proofs,
-            match_atomic_bundle,
+        unimplemented!(
+            "gen_atomic_match_settle_calldata needs to be adapted for IntentOnlyBoundedSettlementBundle"
         )
+    }
+
+    // -----------
+    // | Testing |
+    // -----------
+
+    /// Clear the Merkle tree for testing
+    #[cfg(feature = "integration")]
+    pub async fn clear_merkle_tree(&self) -> Result<TransactionReceipt, DarkpoolClientError> {
+        unimplemented!("clear_merkle_tree not yet implemented for Solidity darkpool")
     }
 }
