+++ crates/workers/task-driver/Cargo.toml
@@ -9,7 +9,7 @@ integration = [
     "types-tasks/mocks",
     "darkpool-client/integration",
     "state/mocks",
-    "test-helpers/arbitrum",
+    "test-helpers/blockchain",
     "util/mocks",
 ]
 
@@ -71,7 +71,7 @@ alloy-primitives = { workspace = true }
 clap = { version = "4.0", features = ["derive"] }
 colored = "2"
 config = { workspace = true }
-darkpool-client = { workspace = true, features = ["arbitrum"] }
+darkpool-client = { workspace = true }
 eyre = { workspace = true }
 flate2 = "1.0"
 inventory = "0.3"
